% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_function.R
\name{inner_function}
\alias{inner_function}
\title{Print the argument and some version info}
\usage{
inner_function(x)
}
\arguments{
\item{x}{The argument to print}
}
\description{
Print the argument and some version info
}
\examples{
inner_function(7)
}
